Remove "Downloadable" from Apple Music
Apple Music is not the same as the iTunes Store (which I believe is
going away soon). Tracks on Apple Music are encrypted and can only be
played on an apple device that's authorized against that account. So
it's not fair to day these are Downloadable.

https://rateyourmusic.com/admin/corq/?req_id=10106042

Also, removed some build commands for individual browsers in the README.

Fixes #148