Quickly add "featured" and "remixer" credits.
These are really common, so having a way to add them with a single click
is a huge time saver.

By selecting any text within the track list the "+ featured" and
"+ remixer" buttons will appear in the "Credits" section. Clicking
either of them will perform the search and selection as well as
including the role and track number.